Move Groovy scripts into their own file
This makes the resource code shorter to read. Also, better editing (and later,
testing?) of the Groovy code.
