+++ files/default/delete_user.groovy
@@ -1,0 +1,3 @@
+import org.sonatype.nexus.security.user.UserManager;
+
+security.securitySystem.deleteUser(args, UserManager.DEFAULT_SOURCE);
